requirements.txt
flask==2.2.5
werkzeug==2.2.3
torch
torchvision
pillow
